[ISSUE-2906] Moving logic to ChannelLogic
