+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/channel/ChannelController.kt
@@ -454,19 +454,7 @@ public class ChannelController internal constructor(
     }
 
     internal suspend fun sendGiphy(message: Message): Result<Message> {
-        val request = SendActionRequest(
-            message.cid,
-            message.id,
-            message.type,
-            mapOf(KEY_MESSAGE_ACTION to MESSAGE_ACTION_SEND)
-        )
-        val result = domainImpl.runAndRetry { channelClient.sendAction(request) }
-        removeLocalMessage(message)
-        return if (result.isSuccess) {
-            Result(result.data())
-        } else {
-            Result(result.error())
-        }
+        return channelLogic.sendGiphy(message)
     }
 
     internal suspend fun shuffleGiphy(message: Message): Result<Message> {
@@ -1059,6 +1047,5 @@ public class ChannelController internal constructor(
     internal companion object {
         private const val KEY_MESSAGE_ACTION = "image_action"
         private const val MESSAGE_ACTION_SHUFFLE = "shuffle"
-        private const val MESSAGE_ACTION_SEND = "send"
     }
 }
